+++ packages/sushi/src/config/token-maps/bases-to-check-trades-against.ts
@@ -508,4 +508,9 @@ export const BASES_TO_CHECK_TRADES_AGAINST: {
     USDC[ChainId.MANTLE],
     METH[ChainId.MANTLE],
   ],
+  [ChainId.CURTIS]: [
+    WNATIVE[ChainId.CURTIS],
+    USDT[ChainId.CURTIS],
+    DAI[ChainId.CURTIS],
+  ],
 }